Sesión06/tema10/ws6-fastapi-jwt-auth/requirements.txt

fastapi
uvicorn
python-jose
pydantic
